packages/projectair/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "projectair"
version = "0.7.0"
description = "Project AIR: forensic reconstruction and incident response for AI agents. Signed AgDR chains, RFC 3161 + Sigstore Rekor anchoring, causal explanation, and Auth0-verified human-in-the-loop containment."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Vindicara", email = "eng@vindicara.io" }]
keywords = ["ai", "security", "incident-response", "forensics", "langchain", "agent", "agdr", "owasp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "blake3>=1.0.0,<2.0",
    # cryptography: floor 46.0.6 patches CVE-2026-26007 (subgroup attack on SECT curves, HIGH)
    # and CVE-2026-34073 (incomplete DNS name constraint, LOW).
    "cryptography>=46.0.6,<47.0",
    "pydantic>=2.7.0,<3.0",
    "typer>=0.12.0,<1.0",
    # langchain-core: floor 1.2.22 patches CVE-2026-34070 (load_prompt path traversal, HIGH)
    # and CVE-2026-26013 (image_url SSRF, LOW).
    "langchain-core>=1.2.22,<2.0",
    # fpdf2: pure-Python PDF rendering for `air trace --format pdf`. No system deps.
    "fpdf2>=2.8.0,<3.0",
    # PyYAML: parses agent registry files for ASI03/ASI10 Zero-Trust enforcement.
    # Load path uses yaml.safe_load only. JSON registries work via stdlib and do
    # not require this package at runtime.
    "pyyaml>=6.0.2,<7.0",
    # rfc3161-client: RFC 3161 trusted timestamping for chain-root anchoring
    # (Layer 1 v1). Carries the ASN.1 parsing and the verifier so the wrapper
    # in airsdk/anchoring/rfc3161.py is integration glue, not custom crypto.
    "rfc3161-client>=1.0,<2.0",
    # sigstore: Rekor transparency-log anchoring. Used for the Hashedrekord
    # entry submission, response parsing, and Merkle inclusion-proof
    # verification. The anchoring path does not call Fulcio; Rekor accepts
    # raw Ed25519 public keys directly.
    "sigstore>=3.0,<5.0",
    # PyJWT[crypto]: Layer 3 step-up approval verification. Validates Auth0
    # (or any OIDC + JWKS) tokens against the tenant's published public keys.
    # The [crypto] extra pulls in cryptography for RS256/RS384/RS512 support.
    "PyJWT[crypto]>=2.8,<3.0",
    # jcs: RFC 8785 JSON Canonicalization Scheme. Layer 4 AgDR Handoff
    # canonicalizes payloads before BLAKE3 hashing so two parties hash
    # the same bytes regardless of dict ordering or whitespace.
    "jcs>=0.2.1,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0",
    "pytest-cov>=5.0.0,<6.0",
    "ruff>=0.6.0,<1.0",
    "mypy>=1.11.0,<2.0",
    "types-PyYAML>=6.0.12,<7.0",
]

[project.scripts]
air = "projectair.cli:main"

[project.urls]
Homepage = "https://vindicara.io"
Repository = "https://github.com/vindicara-inc/projectair"
"Bug Tracker" = "https://github.com/vindicara-inc/projectair/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/airsdk", "src/projectair"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "network: requires outbound network access; skipped by default. Run with `pytest -m network`.",
    "integration: requires external service credentials (e.g. AIR_AUTH0_*). Run with `pytest -m integration`.",
]
addopts = "-m 'not network and not integration'"

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "PT", "SIM", "TCH", "RUF"]
ignore = [
    "S101",   # assert in tests is fine
    "A002",   # shadowing 'input' builtin is intentional SDK API design
    "N818",   # exception names without Error suffix are intentional public API
    "E501",   # long strings in security descriptions are intentional
    "B008",   # typer.Option()/Argument() in defaults is standard idiom, like FastAPI's Depends
    "TC001",  # keep runtime imports for first-party types; readability over optimisation
    "TC002",  # same reasoning for third-party imports used in signatures
    "TC003",  # keep runtime imports for pathlib.Path et al. Typer / pydantic resolve types at runtime.
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
plugins = ["pydantic.mypy"]
